requirement.txt
alembic==1.11.0
async-timeout==4.0.2
blinker==1.6.2
click==8.1.3
coverage==7.2.7
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
passlib==1.7.4
pkg_resources==0.0.0
pluggy==1.2.0
PyJWT==2.7.0
pytest==7.4.0
python-dotenv==1.0.0
redis==4.5.5
rq==1.15.0
six==1.16.0
SQLAlchemy==2.0.13
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.4
zipp==3.15.0
